Aircraft-Manufacturing-App/pyproject.toml
[tool.poetry]
name = "Aircraft-Manufacturing-App"
version = "0.1.0"
description = ""
authors = ["Mina Nilay TEZER <107746643+Mimakk@users.noreply.github.com>"]
readme = "README.rst"
packages = [{ include = "project" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"  
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
psycopg2 = ">=2.9.10,<3.0.0"
django-filter = ">=25.1,<26.0"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"

[poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
